+++ src/utils/num_primitive.rs
@@ -20,6 +20,10 @@ pub(crate) trait PrimitiveFloatFns: Sized {
 
     fn recip(self) -> Self;
 
+    fn max(self, other: Self) -> Self;
+
+    fn min(self, other: Self) -> Self;
+
     fn clamp(self, min: Self, max: Self) -> Self;
 
     fn abs(self) -> Self;
@@ -163,6 +167,16 @@ macro_rules! impl_float {
             #[inline(always)]
             fn recip(self) -> Self {
                 self.recip()
+            }
+
+            #[inline(always)]
+            fn max(self, other: Self) -> Self {
+                self.max(other)
+            }
+
+            #[inline(always)]
+            fn min(self, other: Self) -> Self {
+                self.min(other)
             }
 
             #[inline(always)]